requirements.txt
bcrypt
boto3
flask
pyyaml
